Completed searching and filtering courses based on tag and categories

+++ src/app/api/courses/[courseID]/route.js
@@ -1,7 +1,19 @@
 import { Course } from "@/Models/Course";
+import { UserInfo } from "@/Models/UserInfo";
+import checkIsLoggedIn from "@/middlewares/CheckIsLoggedIn";
+import mongoose from "mongoose";
 
 export async function GET(req, { params }) {
   try {
+    // Checking the user is logged in or not by checking the token;
+    const decoded = checkIsLoggedIn();
+    mongoose.connect(process.env.DATABASE_URL);
+
+    const userInfo = await UserInfo.findOne({ email: decoded.email });
+    if (!userInfo) {
+      throw new Error("You are not authorized!");
+    }
+
     const courseID = params.courseID;
     const result = await Course.findById(courseID);
     if (!result) {
@@ -24,9 +36,22 @@ export async function GET(req, { params }) {
 
 export async function PATCH(req, { params }) {
   try {
+    // Checking the user is logged in or not by checking the token;
+    const decoded = checkIsLoggedIn();
+    mongoose.connect(process.env.DATABASE_URL);
+
+    const userInfo = await UserInfo.findOne({ email: decoded.email });
+    if (!userInfo) {
+      throw new Error("You are not authorized!");
+    }
+
+    if(userInfo.role !== "admin") {
+      throw new Error("You are not authorized to updated course details!");
+    }
+
     const body = await req.json();
-    if(!body) {
-        throw new Error("No data given!");
+    if (!body) {
+      throw new Error("No data given!");
     }
     const courseID = params.courseID;
     const course = await Course.findById(courseID);
@@ -60,9 +85,24 @@ export async function PATCH(req, { params }) {
 
 export async function DELETE(req, { params }) {
   try {
+    // Checking the user is logged in or not by checking the token;
+    const decoded = checkIsLoggedIn();
+    mongoose.connect(process.env.DATABASE_URL);
+
+    const userInfo = await UserInfo.findOne({ email: decoded.email });
+    if (!userInfo) {
+      throw new Error("You are not authorized!");
+    }
+
+    if(userInfo.role !== "admin") {
+      throw new Error("You are not authorized to delete this course!");
+    }
+
+    
     const courseID = params.courseID;
     const result = await Course.findByIdAndDelete(courseID);
     if (!result) {
+      sss;
       throw new Error("Course not found!");
     }
     return Response.json({
